cspm-v2/backend/requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
pydantic==2.6.4
pydantic-core==2.16.3
psycopg2-binary==2.9.9
alembic==1.13.1
boto3==1.34.84
azure-identity==1.15.0
azure-mgmt-compute==30.6.0
azure-mgmt-storage==21.0.0
azure-mgmt-network==25.2.0
azure-mgmt-sql==3.0.1
azure-mgmt-keyvault==10.3.0
google-auth==2.28.2
google-api-python-client==2.123.0
python-dotenv==1.0.1
anyio==4.3.0
starlette==0.37.2
apscheduler==3.10.4
python-multipart==0.0.9
